Push the first set up of gerrit using Kubernetes
This first patch allows to create a persistent gerrit site which saves
the changes. Even if the pod got deleted or restarted because of a
failure the data will be saved through persistent volume.

However many improvementss are still needed.

1- Until now we can only have one pod running the application. If we try
to scale up the application the other pods will fail and throw indexing
errors because of the writing lock on the index hold by the main pod.
This defies Kubernetes principal which allows high-availavility

2- The node needs to be configured manually to be able to create the
directories required to create the persistent volume.

3- Allow the replication of plugins folder to allow custom plugins
inside gerrit application